Make PostSaveEventModel a mixin
